Applied recommendations from review

+++ include/nodes/actuator_nodes/MotorNode.h
@@ -28,7 +28,7 @@ public:
 
     void moveVelocity(float velocity);
 
-    double getVelocity();
+    float getVelocity();
 
     void moveAbsolute(double position, int max_velocity);
