+++ src/nodes/actuator_nodes/MotorNode.cpp
@@ -50,7 +50,7 @@ void MotorNode::moveVelocity(float velocity) {
     m_motor.move_velocity(rpm);
 }
 
-double MotorNode::getVelocity() {
+float MotorNode::getVelocity() {
     float rpm = m_motor.get_actual_velocity();
     return (rpm / m_getMaxRPM()) * MAX_VELOCITY;
 }
